fuel_upgrade_system/fuel_package_updates/requirements.txt
oslo.serialization>=1.0.0
requests>=1.2.3
ordereddict>=1.1
PyYAML>=3.10
python-keystoneclient>=0.11
